Fixes bug when iterating over empty arrays.
The one case I hadn't test was what happens when an empty array is
given in the input, for example:

```text
"header": "Hello!"
"somearray": ["apple", "banana", "citrus"]
"footer": "That's it!"
```

It resulted that as the code was written, it ended in an infinite
loop. This required modifying some of the code to accomodate this
case. On the plus side, some of the code was already in my mind as a
way to implement the keyword `/if`.