Improves performance by avoiding iterating when output won't be shown.
Introduces a small performance issue when rending a loop that won't be
shown. In the past it would iterate over all elements of the array
replacing every tag as if were being shown.

With this change it treats elements what won't be shown as if it were
dealing with an empty array.